LibPythonAI/python_ai_lib/requirements.txt
SQLAlchemy
psycopg-binary
psutil
tiktoken
chardet

openai

langchain==0.3.4
langchain-openai==0.2.3
langchain-community==0.3.3
langchain-chroma
langchain_postgres

chroma-hnswlib==0.7.3
chromadb==0.5.0

magika==0.6.0rc3
pillow
python-docx
python-dotenv
python-pptx
openpyxl
markdown
pdfminer.six

autogen-agentchat
autogen-ext
beautifulsoup4
wikipedia
flaml[automl]
selenium
webdriver-manager
duckduckgo-search

# flask-socketio
# Flask-Cors
aiohttp
python-socketio

arxiv

# promptflow-core
promptflow-devkit
# promptflow-tracing

